combined 2d and 3d choropleths
