+++ python/cuxfilter/charts/deckgl/__init__.py
@@ -1,1 +1,2 @@
 from .bindings import PolygonDeckGL, TS_CODE
+from .deckgl import choropleth
